🐛 Fix: fix code

+++ src/main/java/com/words_hanjoom/domain/wordbooks/entity/Word.java
@@ -18,7 +18,7 @@ public class Word {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "word_id")
-    private Integer wordId;
+    private Long wordId;
 
     @Column(name = "word_name", length = 100, nullable = false)
     private String wordName;
